refactor: 🚚 organized the `/public` folder
Related: #179
