+++ src/components/Team/team-config.ts
@@ -1,13 +1,13 @@
 import { type StaticImageData } from 'next/image';
 
-import Aditya_Kulkarni from '@/public/Team/Executives/Aditya Kulkarni.jpg';
-import AnaqiAmir from '@/public/Team/Executives/Anaqi Amir.jpg';
-import ChristineVan from '@/public/Team/Executives/Christine Van.png';
-import IvyChen from '@/public/Team/Executives/Ivy Chen.jpg';
-import KhadeemSankar from '@/public/Team/Executives/Khadeem Sankar.jpg';
-import MalharPandya from '@/public/Team/Executives/Malhar Pandya.jpg';
-import ThanareeSrirawewongsa from '@/public/Team/Executives/Thanaree Srirawewongsa (Jean).jpg';
-import TiannePane from '@/public/Team/Executives/Tianne Pane.jpg';
+import Aditya_Kulkarni from '@/public/Main/Executives/Aditya Kulkarni.jpg';
+import AnaqiAmir from '@/public/Main/Executives/Anaqi Amir.jpg';
+import ChristineVan from '@/public/Main/Executives/Christine Van.png';
+import IvyChen from '@/public/Main/Executives/Ivy Chen.jpg';
+import KhadeemSankar from '@/public/Main/Executives/Khadeem Sankar.jpg';
+import MalharPandya from '@/public/Main/Executives/Malhar Pandya.jpg';
+import ThanareeSrirawewongsa from '@/public/Main/Executives/Thanaree Srirawewongsa (Jean).jpg';
+import TiannePane from '@/public/Main/Executives/Tianne Pane.jpg';
 
 export interface TeamMember {
   name: string;
